pyproject.toml
[project]
name = "flask_app"
version = "0.1.0"
description = ""
authors = [
    {name = "ruslana-tymchyk",email = "tymchyk3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "firebase-admin (>=6.6.0,<7.0.0)",
    "groq (>=0.16.0,<0.17.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flask-limiter (>=3.12,<4.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "openai (==1.75.0)",
    "supabase (>=2.15.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
